Move the radio option into its own component `SurveyRadioOption`
