+++ client/components/survey-container/components/question-single-selection.tsx
@@ -1,56 +1,18 @@
 import { QuestionSelectionType } from './question-step';
+import SurveyRadioOption from './survey-radio-option';
 
 const QuestionSingleSelection = ( { onChange, question, value }: QuestionSelectionType ) => {
 	return (
 		<div className="question-options__container" role="radiogroup">
-			{ question.options.map( ( option, index ) => {
-				const isSelected = value.includes( option.value );
-				const handleKeyDown = ( event: React.KeyboardEvent< HTMLDivElement > ) => {
-					if ( event.key === 'Enter' || event.key === ' ' ) {
-						onChange( question.key, [ option.value ] );
-					}
-				};
-
-				return (
-					<div
-						key={ index }
-						className={ `question-options__option-control components-radio-control__option ${
-							isSelected ? 'checked' : ''
-						}` }
-						role="radio"
-						tabIndex={ 0 }
-						onClick={ () => onChange( question.key, [ option.value ] ) }
-						onKeyDown={ handleKeyDown }
-						aria-checked={ isSelected ? 'true' : 'false' }
-						aria-labelledby={ `option-label-${ option.value } option-help-text-${ option.value }` }
-					>
-						<input
-							type="radio"
-							id={ `option-${ option.value }` }
-							name={ question.key }
-							value={ option.value }
-							onChange={ () => onChange( question.key, [ option.value ] ) }
-							checked={ isSelected }
-							className="form-radio"
-							tabIndex={ -1 }
-							aria-hidden="true"
-						/>
-						<div className="question-options__option-label">
-							<label id={ `option-label-${ option.value }` } htmlFor={ `option-${ option.value }` }>
-								{ option.label }
-							</label>
-							{ option.helpText && (
-								<span
-									id={ `option-help-text-${ option.value }` }
-									className="question-options__option-help-text"
-								>
-									{ option.helpText }
-								</span>
-							) }
-						</div>
-					</div>
-				);
-			} ) }
+			{ question.options.map( ( option, index ) => (
+				<SurveyRadioOption
+					key={ index }
+					question={ question }
+					option={ option }
+					onChange={ onChange }
+					value={ value }
+				></SurveyRadioOption>
+			) ) }
 		</div>
 	);
 };
